add article column

+++ article/views.py
@@ -4,30 +4,48 @@ from .models import ArticlePost
 import markdown
 from .forms import ArticlePostForm
 from django.contrib.auth.models import User
+from django.core.paginator import Paginator
+from django.contrib.auth.decorators import login_required
+from .models import ArticleColumn
 
 # Create your views here.
 def article_list(request):
-    articles = ArticlePost.objects.all()
-    context = {'articles': articles}
+    # 修改变量名称（articles -> article_list）
+    article_list = ArticlePost.objects.all()
+
+    # 每页显示 6 篇文章
+    paginator = Paginator(article_list, 6)
+    # 获取 url 中的页码
+    page = request.GET.get('page')
+    # 将导航对象相应的页码内容返回给 articles
+    articles = paginator.get_page(page)
+
+    context = { 'articles': articles }
     return render(request, 'article/list.html', context)
 
 # 文章详情
 def article_detail(request, id):
     article = ArticlePost.objects.get(id=id)
-
-    # 将markdown语法渲染成html样式
-    article.body = markdown.markdown(article.body,
+    # 浏览量 +1
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
+    # 修改 Markdown 语法渲染
+    md = markdown.Markdown(
         extensions=[
-        # 包含 缩写、表格等常用扩展
         'markdown.extensions.extra',
-        # 语法高亮扩展
         'markdown.extensions.codehilite',
-        ])
+        'markdown.extensions.toc',
+        ]
+    )
+    article.body = md.convert(article.body)
 
-    context = { 'article': article }
+    # 新增了md.toc对象
+    context = { 'article': article, 'toc': md.toc }
     return render(request, 'article/detail.html', context)
 
 # 写文章的视图
+# 检查登录
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     # 判断用户是否提交数据
     if request.method == "POST":
@@ -40,7 +58,9 @@ def article_create(request):
             # 指定数据库中 id=1 的用户为作者
             # 如果你进行过删除数据表的操作，可能会找不到id=1的用户
             # 此时请重新创建用户，并传入此用户的id
-            new_article.author = User.objects.get(id=1)
+            new_article.author = User.objects.get(id=request.user.id)
+            if request.POST['column'] != 'none':
+                new_article.column = ArticleColumn.objects.get(id=request.POST['column'])
             # 将新文章保存到数据库中
             new_article.save()
             # 完成后返回到文章列表
@@ -53,12 +73,19 @@ def article_create(request):
         # 创建表单类实例
         article_post_form = ArticlePostForm()
         # 赋值上下文
-        context = { 'article_post_form': article_post_form }
+        # 新增及修改的代码
+        columns = ArticleColumn.objects.all()
+        context = { 'article_post_form': article_post_form, 'columns': columns }
         # 返回模板
         return render(request, 'article/create.html', context)
 
 # 安全删除文章
+# 提醒用户登录
+@login_required(login_url='/userprofile/login/')
 def article_safe_delete(request, id):
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     if request.method == 'POST':
         article = ArticlePost.objects.get(id=id)
         article.delete()
@@ -67,6 +94,8 @@ def article_safe_delete(request, id):
         return HttpResponse("仅允许post请求")
 
 # 更新文章
+# 提醒用户登录
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     """
     更新文章的视图函数
@@ -77,6 +106,9 @@ def article_update(request, id):
 
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     # 判断用户是否为 POST 提交表单数据
     if request.method == "POST":
         # 将提交的数据赋值到表单实例中
@@ -86,6 +118,11 @@ def article_update(request, id):
             # 保存新写入的 title、body 数据并保存
             article.title = request.POST['title']
             article.body = request.POST['body']
+            # 新增的代码
+            if request.POST['column'] != 'none':
+                article.column = ArticleColumn.objects.get(id=request.POST['column'])
+            else:
+                article.column = None
             article.save()
             # 完成后返回到修改后的文章中。需传入文章的 id 值
             return redirect("article:article_detail", id=id)
@@ -98,6 +135,11 @@ def article_update(request, id):
         # 创建表单类实例
         article_post_form = ArticlePostForm()
         # 赋值上下文，将 article 文章对象也传递进去，以便提取旧的内容
-        context = { 'article': article, 'article_post_form': article_post_form }
+        columns = ArticleColumn.objects.all()
+        context = { 
+            'article': article, 
+            'article_post_form': article_post_form,
+            'columns': columns,
+        }
         # 将响应返回到模板中
         return render(request, 'article/update.html', context)
